fix: #3136 geo diary fetching improvements
