+++ packages/geo-diary/src/components/pages/appointment/appointment.tsx
@@ -1,17 +1,16 @@
 import * as React from 'react'
 import { useQuery } from '@apollo/react-hooks'
-import { History } from 'history'
 import dayjs from 'dayjs'
 import qs from 'query-string'
 import { Loader, isMobile } from '@reapit/elements'
 import { ExtendedAppointmentModel } from '@/types/global'
 import GET_APPOINTMENTS from './get-appointments.graphql'
 import { MobileLayout } from './mobile-layout'
-import { useHistory, useLocation } from 'react-router-dom'
+import { useLocation } from 'react-router-dom'
 import { DesktopLayout } from './desktop-layout'
-import { ROUTES } from '@/core/router'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
 import { useReapitConnect } from '@reapit/connect-session'
+import { memo } from 'react'
 
 export type AppointmentProps = {}
 
@@ -84,33 +83,15 @@ export const sortAppoinmentsByStartTime = (
   return sortedAppoinments
 }
 
-export type HandleUseEffectParams = {
-  queryParams: qs.ParsedQuery<string>
-  history: History
-}
-
-export const handleUseEffect = ({ queryParams, history }: HandleUseEffectParams) => () => {
-  if (navigator.geolocation) {
-    navigator.geolocation.getCurrentPosition(position => {
-      const queryString = qs.stringify({
-        ...queryParams,
-        currentLat: position.coords.latitude,
-        currentLng: position.coords.longitude,
-      })
-      history.push(`${ROUTES.APPOINTMENT}?${queryString}`)
-    })
-  }
-}
-
 export const Appointment: React.FC<AppointmentProps> = () => {
   const isMobileView = isMobile()
-  const session = useReapitConnect(reapitConnectBrowserSession)
-  const userCode = session.connectSession?.loginIdentity.userCode || ''
+  const { connectSession } = useReapitConnect(reapitConnectBrowserSession)
+  const userCode = connectSession?.loginIdentity.userCode ?? ''
   const location = useLocation()
-  const history = useHistory()
   const queryParams = qs.parse(location.search)
   const time = queryParams.time || 'today'
   const { start, end } = startAndEndTime[time]
+
   const { data, loading } = useQuery<AppointmentListQueryData, AppointmentListQueryVariables>(GET_APPOINTMENTS, {
     variables: {
       negotiatorId: [userCode],
@@ -119,8 +100,9 @@ export const Appointment: React.FC<AppointmentProps> = () => {
       includeCancelled: true,
       includeConfirm: true,
     },
+    skip: !userCode,
   })
-  React.useEffect(handleUseEffect({ queryParams, history }), [])
+
   const appointmentSorted = React.useMemo(sortAppoinmentsByStartTime(data?.GetAppointments?._embedded || []), [
     data?.GetAppointments?._embedded,
   ])
@@ -133,4 +115,4 @@ export const Appointment: React.FC<AppointmentProps> = () => {
   return <DesktopLayout appointments={appointmentSorted} />
 }
 
-export default React.memo(Appointment)
+export default memo(Appointment)
