+++ packages/geo-diary/src/core/private-route-wrapper.tsx
@@ -2,20 +2,46 @@ import * as React from 'react'
 import Menu from '@/components/ui/menu'
 import { Loader, AppNavContainer, Helper } from '@reapit/elements'
 import ErrorBoundary from './error-boundary'
-import { Redirect, useLocation } from 'react-router'
+import { Redirect, useHistory, useLocation } from 'react-router'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
 import { useReapitConnect } from '@reapit/connect-session'
 import { ApolloProvider } from '@apollo/react-hooks'
 import getClient from '@/graphql/client'
+import qs from 'query-string'
+import { ROUTES } from '@/core/router'
+import { History } from 'history'
+import { useEffect } from 'react'
 
 const { Suspense } = React
 
 export type PrivateRouteWrapperProps = {}
 
+export type HandleUseEffectParams = {
+  queryParams: qs.ParsedQuery<string>
+  history: History
+}
+
+export const handleUseEffect = ({ queryParams, history }: HandleUseEffectParams) => () => {
+  if (navigator.geolocation) {
+    navigator.geolocation.getCurrentPosition(position => {
+      const queryString = qs.stringify({
+        ...queryParams,
+        currentLat: position.coords.latitude,
+        currentLng: position.coords.longitude,
+      })
+      history.push(`${ROUTES.APPOINTMENT}?${queryString}`)
+    })
+  }
+}
+
 export const PrivateRouteWrapper: React.FC<PrivateRouteWrapperProps> = ({ children }) => {
   const { connectSession, connectInternalRedirect } = useReapitConnect(reapitConnectBrowserSession)
   const location = useLocation()
   const currentUri = `${location.pathname}${location.search}`
+  const history = useHistory()
+  const queryParams = qs.parse(location.search)
+
+  useEffect(handleUseEffect({ queryParams, history }), [queryParams])
 
   if (!connectSession) {
     return (
